Restructure `batterysync/wear`
