+++ batterysync/wear/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/data/BatterySyncStateDsRepository.kt
@@ -1,9 +1,11 @@
-package com.boswelja.smartwatchextensions.batterysync
+package com.boswelja.smartwatchextensions.batterysync.data
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.Serializer
 import androidx.datastore.dataStore
+import com.boswelja.smartwatchextensions.batterysync.domain.BatterySyncState
+import com.boswelja.smartwatchextensions.batterysync.domain.BatterySyncStateRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.decodeFromByteArray
