+++ batterysync/wear/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/di/DI.kt
@@ -1,6 +1,11 @@
 package com.boswelja.smartwatchextensions.batterysync
 
 import androidx.core.content.getSystemService
+import com.boswelja.smartwatchextensions.batterysync.data.BatteryStatsDsRepository
+import com.boswelja.smartwatchextensions.batterysync.data.BatterySyncStateDsRepository
+import com.boswelja.smartwatchextensions.batterysync.domain.BatteryStatsRepository
+import com.boswelja.smartwatchextensions.batterysync.domain.BatterySyncStateRepository
+import com.boswelja.smartwatchextensions.batterysync.platform.WearBatterySyncNotificationHandler
 import com.boswelja.smartwatchextensions.batterysync.ui.BatteryStatsViewModel
 import com.boswelja.smartwatchextensions.core.devicemanagement.phoneStateStore
 import org.koin.android.ext.koin.androidContext
