+++ batterysync/wear/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/domain/BatteryStatsRepository.kt
@@ -1,5 +1,6 @@
-package com.boswelja.smartwatchextensions.batterysync
+package com.boswelja.smartwatchextensions.batterysync.domain
 
+import com.boswelja.smartwatchextensions.batterysync.BatteryStats
 import kotlinx.coroutines.flow.Flow
 
 /**
